fix order bug

+++ src/modules/categories/categories.service.ts
@@ -1,15 +1,24 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ConflictException,
+  Inject,
+  Injectable,
+  NotFoundException,
+  forwardRef,
+} from '@nestjs/common';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Category } from './entities/category.entity';
 import { Repository } from 'typeorm';
+import { ProductsService } from '../products/products.service';
 
 @Injectable()
 export class CategoriesService {
   constructor(
     @InjectRepository(Category)
     private readonly categoryRepo: Repository<Category>,
+    @Inject(forwardRef(() => ProductsService))
+    private readonly productService: ProductsService,
   ) {}
   create(createCategoryDto: CreateCategoryDto) {
     return this.categoryRepo.save(createCategoryDto);
@@ -25,13 +34,17 @@ export class CategoriesService {
 
   async update(id: number, updateCategoryDto: UpdateCategoryDto) {
     const category = await this.findOne(id);
-
-    if (!category) throw new NotFoundException();
-
+    if (!category) throw new NotFoundException('category not found');
     return this.categoryRepo.save({ ...category, ...updateCategoryDto });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const isFound = await this.productService.findOne(null, id);
+    if (isFound)
+      throw new ConflictException(
+        'This category cannot be deleted because it is linked with some products ',
+      );
+
     return this.categoryRepo.delete({ id });
   }
 }
